Add files via upload
The dataset featured in the repository includes detailed records on the causes of death in the United States spanning from 2000 to 2019, broken down by age and gender. It's designed to facilitate an in-depth analysis of mortality patterns, spotlighting the role of non-communicable diseases and their varying effects across different demographic groups. This collection of data is crucial for crafting informed public health strategies and for advancing targeted healthcare interventions.